file1.c and file2.c fixed for test scripts
