+++ example_files/file1.c
@@ -2,15 +2,14 @@
 //asdf
 
 #include "PID.h"
-#include "otherfile"
 
 
 
-typedef structtypedefC
+
+typedef structtypedef
 {
 	float derivState;      	/* Last position input
 	*/
-        float NEW_VARIABLE
 	float intgrtState;      	/* Integrator state */
 	float intgrtMax;    /* Maximum allowable integrator state   */
 	float intgrtMin;    /* Minimum allowable integrator state   */
@@ -19,6 +18,9 @@ typedef structtypedefC
 	float kP;    	/* proportional gain    */
 	float kD;     	/* derivative gain  */
 } SPid;
+//comment
+
+
 
 void InitPID(Spid* pid, float kP, float kI, float kD, float iMin, float iMax)
 {
@@ -28,6 +30,7 @@ void InitPID(Spid* pid, float kP, float kI, float kD, float iMin, float iMax)
 	pid->intgrtMin = iMin;
 	pid->intgrtMax = iMax;
 }
+//other comment
 
 void InitPID(Spid* pid, float kP, float kI, float kD, float iMin, float iMax)
 {
@@ -37,17 +40,6 @@ void InitPID(Spid* pid, float kP, float kI, float kD, float iMin, float iMax)
 	pid->intgrtMin = iMin;
 	pid->intgrtMax = iMax;
 }
-
-
-}
-void InitPIDfunctionB(Spid* pid, float kP, float kI, float kD, float iMin, float iMax)
-{
-	pid->kP = kP;
-	pid->kI = kI;
-	pid->k_D = kD;
-	pid->intgrtMin = iMin;
-	pid->intgrtMax = iMax;
-}
-//changed comment
+//other comment
 
 }
